requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.6.5
numpy==1.21.4
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.12.0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
shortuuid==1.0.8
six==1.16.0
stemming==1.0.1
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
